Site: retire the early-access CTA, surface the terminal installs
The nav kept a scroll-gated 'Get early access' mailto button and the
hero download button flashed the same copy while /latest.json loaded —
both leftovers from before the first release. Every CTA is now a
download: a shared latest.ts composable links each one to the GitHub
releases page from first paint and swaps in the direct DMG link once
the manifest resolves, so no placeholder copy ever shows.

The Homebrew command moves up next to the hero buttons, and a new
curl | bash path joins it: site/public/install.sh is served at
keelhaven.app/install.sh and installs the latest release DMG with
nothing beyond what ships with macOS. README and FAQ carry both.
