+++ docs/WEBSITE.md
@@ -20,7 +20,9 @@ Content lives in `site/index.md` (the landing page — all copy sits in its
 frontmatter so a future `site/zh/index.md` can localize it in one place),
 `site/privacy.md`, `site/licenses.md`, and the components under
 `site/.vitepress/theme/`. Images in `site/public/` are generated by
-`design/build.mjs` — don't hand-edit them.
+`design/build.mjs` — don't hand-edit them. `site/public/install.sh` is the
+exception: it is source, served verbatim at `keelhaven.app/install.sh` for the
+`curl | bash` install path the landing page advertises.
 
 ## How it gets published
